refactor: move column processing to useProcessedColumns hook
